Add --typecheck flag for Setup build
This patch adds a --typecheck flag to the build command in
Distribution.Simple.Setup. This flag is plumbed through to the build functions
in Distribuion.Simple.GHC where it enables the ghcOptNoCode flag, which causes
-fno-code and -fwrite-interface to be passed to ghc.

The logic in Distribution.Simple.GHC.buildOrReplLib already considered this
case, but only for indefinite components. The logic in
Distribution.Simple.GHC.gbuild is new.

The flag is not currently passed through to GHCJS, though it could easily be.

There is currently no interface for cabal-install to pass the flag.